# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Este es el proyecto final de la materia espcialización backend 3 para crear, editar, consultar y borrar las entidades de paciente, dentista y turnos.",
        "title": "Proyecto Final Especialización backend 3",
        "contact": {
            "name": "API Support (Natalia Garcia, Federico Bustamante, Damian, Camilo Zuleta)"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http:www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/appointments": {
            "post": {
                "description": "Creates an Appointment, you must send the fields required to process your request Patient, Dentist, Date, Description",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Appointment"
                ],
                "summary": "Creates an Appointment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "publicKey",
                        "name": "PUBLIC-KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "secretKey",
                        "name": "SECRET_KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Create Appointment",
                        "name": "Appointment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/appointments.Appointment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/appointments.Appointment"
                        }
                    }
                }
            }
        },
        "/appointments/{id}": {
            "get": {
                "description": "Gets all appointments if any by patient dni",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Appointment"
                ],
                "summary": "Gets all appointments by dni",
                "parameters": [
                    {
                        "type": "string",
                        "description": "publicKey",
                        "name": "PUBLIC-KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "secretKey",
                        "name": "SECRET_KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "DNI",
                        "name": "DNI",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/appointments.Appointment"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an Appointment by id, you must send all of the appointment fields to process your request",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Appointment"
                ],
                "summary": "Updates an Appointment by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "publicKey",
                        "name": "PUBLIC-KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "secretKey",
                        "name": "SECRET_KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Appointment",
                        "name": "Appointment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/appointments.Appointment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/appointments.Appointment"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes an Appointment by id, be careful with this action.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Appointment"
                ],
                "summary": "Deletes an Appointment by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "publicKey",
                        "name": "PUBLIC-KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "secretKey",
                        "name": "SECRET_KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Delete Appointment",
                        "name": "Appointment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/appointments.Appointment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/appointments.Appointment"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates an Appointment by id, you can send only the appointment fields you need to change",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Appointment"
                ],
                "summary": "Updates an Appointment by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "publicKey",
                        "name": "PUBLIC-KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "secretKey",
                        "name": "SECRET_KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Appointment",
                        "name": "Appointment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/appointments.Appointment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/appointments.Appointment"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "appointments.Appointment": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "dentist": {
                    "$ref": "#/definitions/dentists.Dentist"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "patient": {
                    "$ref": "#/definitions/patients.Patient"
                }
            }
        },
        "dentists.Dentist": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "license": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "patients.Patient": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "dni": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "registration_date": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        }
    }
}